lint(lib): Path-Glueing elimination
#13, #22, #29

+++ lib/shortcuts.ps1
@@ -1,5 +1,5 @@
-'core', 'manifest' | ForEach-Object {
-    . "$PSScriptRoot\$_.ps1"
+'core', 'Helpers', 'manifest' | ForEach-Object {
+    . (Join-Path $PSScriptRoot "$_.ps1")
 }
 
 # Creates shortcut for the app in the start menu
@@ -24,20 +24,22 @@ function create_startmenu_shortcuts($manifest, $dir, $global, $arch) {
 }
 
 function shortcut_folder($global) {
-    $directory = [System.IO.Path]::Combine([Environment]::GetFolderPath('startmenu'), 'Programs', 'Scoop Apps')
-    if ($global) {
-        $directory = [System.IO.Path]::Combine([Environment]::GetFolderPath('commonstartmenu'), 'Programs', 'Scoop Apps')
-    }
-    return $(ensure $directory)
+    $base = if ($global) { 'commonstartmenu' } else { 'startmenu' }
+    $directory = [System.Environment]::GetFolderPath($base) | Join-Path -ChildPath 'Programs\Scoop Apps'
+
+    return ensure $directory
 }
 
 function startmenu_shortcut([System.IO.FileInfo] $target, $shortcutName, $arguments, [System.IO.FileInfo]$icon, $global) {
+    $base = "Creating shortcut for $shortcutName ($(fname $target))"
+    $baseError = "$base failed:"
+
     if (!$target.Exists) {
-        Write-Host -f DarkRed "Creating shortcut for $shortcutName ($(fname $target)) failed: Couldn't find $target"
+        Write-UserMessage -Message "$baseError Couldn't find $target" -Color DarkRed
         return
     }
     if ($icon -and !$icon.Exists) {
-        Write-Host -f DarkRed "Creating shortcut for $shortcutName ($(fname $target)) failed: Couldn't find icon $icon"
+        Write-UserMessage -Message "$baseError Couldn't find icon $icon" -Color DarkRed
         return
     }
 
@@ -48,41 +50,24 @@ function startmenu_shortcut([System.IO.FileInfo] $target, $shortcutName, $argume
     }
 
     $wsShell = New-Object -ComObject WScript.Shell
-    $wsShell = $wsShell.CreateShortcut("$scoop_startmenu_folder\$shortcutName.lnk")
+    $wsShell = $wsShell.CreateShortcut((Join-Path $scoop_startmenu_folder "$shortcutName.lnk"))
     $wsShell.TargetPath = $target.FullName
     $wsShell.WorkingDirectory = $target.DirectoryName
-    if ($arguments) {
-        $wsShell.Arguments = $arguments
-    }
-    if ($icon -and $icon.Exists) {
-        $wsShell.IconLocation = $icon.FullName
-    }
+    if ($arguments) { $wsShell.Arguments = $arguments }
+    if ($icon -and $icon.Exists) { $wsShell.IconLocation = $icon.FullName }
     $wsShell.Save()
-    write-host "Creating shortcut for $shortcutName ($(fname $target))"
+
+    Write-UserMessage -Message $base
 }
 
 # Removes the Startmenu shortcut if it exists
 function rm_startmenu_shortcuts($manifest, $global, $arch) {
-    $shortcuts = @(arch_specific 'shortcuts' $manifest $arch)
-    $shortcuts | Where-Object { $null -ne $_ } | ForEach-Object {
+    @(arch_specific 'shortcuts' $manifest $arch) | Where-Object { $null -ne $_ } | ForEach-Object {
         $name = $_.item(1)
-        $shortcut = "$(shortcut_folder $global)\$name.lnk"
-        write-host "Removing shortcut $(friendly_path $shortcut)"
-        if (Test-Path -Path $shortcut) {
-            Remove-Item $shortcut
-        }
-        # Before issue 1514 Startmenu shortcut removal
-        #
-        # Shortcuts that should have been installed globally would
-        # have been installed locally up until 27 June 2017.
-        #
-        # TODO: Remove this 'if' block and comment after
-        #       27 June 2018.
-        if ($global) {
-            $shortcut = "$(shortcut_folder $false)\$name.lnk"
-            if (Test-Path -Path $shortcut) {
-                Remove-Item $shortcut
-            }
-        }
+        $shortcut = shortcut_folder $global | Join-Path -ChildPath "$name.lnk"
+
+        Write-UserMessage -Message "Removing shortcut $(friendly_path $shortcut)"
+
+        if (Test-Path $shortcut) { Remove-Item $shortcut }
     }
 }
